move memory logging into separate module
